+++ src/main/java/com/example/foodthought/repository/FollowRepository.java
@@ -1,4 +1,4 @@
-package com.example.foodthought.repository.follow;
+package com.example.foodthought.repository;
 
 import com.example.foodthought.entity.Follow;
 import org.springframework.data.jpa.repository.JpaRepository;
